moved code from app to peopleTable
