+++ src/App.tsx
@@ -1,7 +1,5 @@
 import {
   FC,
-  useEffect,
-  useState,
 } from 'react';
 
 import {
@@ -9,7 +7,6 @@ import {
   Routes,
   Route,
   Navigate,
-  useMatch,
 } from 'react-router-dom';
 
 import cn from 'classnames';
@@ -18,121 +15,100 @@ import './App.scss';
 
 import { PeopleTable } from './components/PeopleTable';
 
-export const App: FC = () => {
-  const [selectedSlug, setSelectedSlug] = useState('');
+export const App: FC = () => (
+  <div data-cy="app">
+    <nav
+      data-cy="nav"
+      className="navbar is-fixed-top has-shadow"
+      role="navigation"
+      aria-label="main navigation"
+    >
+      <div className="container">
+        <div className="navbar-brand">
+          <NavLink
+            className={({ isActive }) => (
+              cn(
+                'navbar-item',
+                {
+                  'has-background-grey-lighter': isActive,
+                },
+              )
+            )}
+            to="/"
+          >
+            Home
+          </NavLink>
 
-  const handlePersonClick = (slug: string) => {
-    setSelectedSlug(slug);
-  };
-
-  const match = useMatch('/people/:selectedSlug');
-
-  useEffect(() => {
-    setSelectedSlug(match?.params.selectedSlug || '');
-  }, [match]);
-
-  return (
-
-    <div data-cy="app">
-      <nav
-        data-cy="nav"
-        className="navbar is-fixed-top has-shadow"
-        role="navigation"
-        aria-label="main navigation"
-      >
-        <div className="container">
-          <div className="navbar-brand">
-            <NavLink
-              className={({ isActive }) => (
-                cn(
-                  'navbar-item',
-                  {
-                    'has-background-grey-lighter': isActive,
-                  },
-                )
-              )}
-              to="/"
-            >
-              Home
-            </NavLink>
-
-            <NavLink
-              className={({ isActive }) => (
-                cn(
-                  'navbar-item',
-                  {
-                    'has-background-grey-lighter': isActive,
-                  },
-                )
-              )}
-              to="/people"
-            >
-              People
-            </NavLink>
-          </div>
+          <NavLink
+            className={({ isActive }) => (
+              cn(
+                'navbar-item',
+                {
+                  'has-background-grey-lighter': isActive,
+                },
+              )
+            )}
+            to="/people"
+          >
+            People
+          </NavLink>
         </div>
-      </nav>
+      </div>
+    </nav>
 
-      <main className="section">
-        <div className="container">
-          <Routes>
-            <Route
-              path="/"
-              element={
-                <h1 className="title">Home Page</h1>
-              }
-            />
+    <main className="section">
+      <div className="container">
+        <Routes>
+          <Route
+            path="/"
+            element={
+              <h1 className="title">Home Page</h1>
+            }
+          />
 
-            <Route
-              path="/home"
-              element={
-                <Navigate to="/" replace />
-              }
-            />
+          <Route
+            path="/home"
+            element={
+              <Navigate to="/" replace />
+            }
+          />
 
-            <Route
-              path="/people"
-              element={(
-                <>
-                  <h1 className="title">People Page</h1>
-                  <div className="block">
-                    <div className="box table-container">
-                      <PeopleTable
-                        handlePersonClick={handlePersonClick}
-                        selectedSlug={selectedSlug}
-                      />
-                    </div>
+          <Route
+            path="/people"
+            element={(
+              <>
+                <h1 className="title">People Page</h1>
+                <div className="block">
+                  <div className="box table-container">
+                    <PeopleTable />
                   </div>
-                </>
-              )}
-            />
+                </div>
+              </>
+            )}
+          />
 
-            <Route
-              path="/people/:selectedSlug"
-              element={(
-                <>
-                  <h1 className="title">People Page</h1>
-                  <div className="block">
-                    <div className="box table-container">
-                      <PeopleTable
-                        handlePersonClick={handlePersonClick}
-                        selectedSlug={selectedSlug}
-                      />
-                    </div>
+          <Route
+            path="/people/:selectedSlug"
+            element={(
+              <>
+                <h1 className="title">People Page</h1>
+                <div className="block">
+                  <div className="box table-container">
+                    <PeopleTable />
                   </div>
-                </>
-              )}
-            />
+                </div>
+              </>
+            )}
+          />
 
-            <Route
-              path="*"
-              element={
-                <h1 className="title">Page not found</h1>
-              }
-            />
-          </Routes>
-        </div>
-      </main>
-    </div>
-  );
-};
+          <Route
+            path="*"
+            element={
+              <h1 className="title">Page not found</h1>
+            }
+          />
+        </Routes>
+      </div>
+    </main>
+  </div>
+);
